requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.2.0
dj-database-url==2.1.0
Django==4.2.4
django-channels==0.7.0
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
Markdown==3.4.4
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
pdfkit==1.0.0
Pillow==10.0.0
psycopg==3.1.10
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
razorpay==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
twilio==8.7.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2